search-results page, which would contain results for both badges and pathways. tried to follow the same pattern of the other pages in terms of models, collections, a views. also now including the directory client
