+++ clientapp/models/badge-class.js
@@ -2,7 +2,6 @@ var HumanModel = require('human-model');
 
 module.exports = HumanModel.define({
   type: 'badge_class',
-  url: '/api/badges',
   props: {
     location: ['string'],
     name: ['string'],
