+++ clientapp/models/badge-classes.js
@@ -2,6 +2,5 @@ var BadgeClass = require('./badge-class'),
     Backbone   = require('backbone');
 
 module.exports = Backbone.Collection.extend({
-  model: BadgeClass,
-  url: '/api/badges'
+  model: BadgeClass
 });
