+++ clientapp/templates/includes/search-result.html
@@ -1,7 +1,24 @@
-<section class="listing">
-  <div class="row js-items">
+<div class="small-12 medium-4 large-3 columns">
+  <div class="achievement">
+    <div class="token">
+      <div>
+        <a href="#" class="item-link js-view-item">
+          <img src="{{image}}"/>
+        </a>
+        <a href="#" class="wishlist-button js-toggle-wishlist">
+          <i class="fa fa-heart favorite-icon js-favorite-icon"></i>
+        </a>
+      </div>
+    </div>
+    <div class="data">
+      <h5>{{name}}</h5>
+      {% for tag in tags %}
+      <span class="secondary radius label">#{{tag}}</span>
+      {% endfor %}
+      <div>
+        <img src="/static/sbux.jpg" width=50>
+        Created by...
+      </div>
+    </div>
   </div>
-  <div class="text-center">
-    <a class="js-view-more" href="#">View More</a>
-  </div>
-</section>
+</div>
