+++ clientapp/templates/includes/search-results.html
@@ -1,0 +1,7 @@
+<section class="listing">
+  <div class="row js-items">
+  </div>
+  <div class="text-center">
+    <a class="js-view-more" href="#">View More</a>
+  </div>
+</section>
