+++ clientapp/views/includes/search-result.js
@@ -1,15 +1,12 @@
 var HumanView = require('human-view');
 var templates = require('templates');
-var AchievementView = require('./achievement');
+var util = require('util');
 
 module.exports = HumanView.extend({
   template: templates.includes.searchResult,
-  render: function () {
-    this.renderAndBind({});
-    if (this.collection.length) {
-      this.renderCollection(this.collection, AchievementView, this.$('.js-items')[0]);
-    }
-    this.collection.on('sync', this.render.bind(this));
-    return this;
+  render: function (opts) {
+    var container = opts.containerEl;
+    this.renderAndBind(this.model);
+    this.$el.appendTo(container);
   }
 });
