+++ clientapp/views/pages/search.js
@@ -1,14 +1,14 @@
 var HumanView = require('human-view');
 var templates = require('templates');
-var SearchResult = require('../includes/search-result');
+var SearchResults = require('../includes/search-results');
 var Backbone = require('backbone');
 
 module.exports = HumanView.extend({
   template: templates.pages.search,
   render: function () {
     this.renderAndBind({});//this.model);
-    this.renderSubview(new SearchResult({
-      collection: new Backbone.Collection(this.model.badges)
+    this.renderSubview(new SearchResults({
+      collection: this.model.badges
     }), '.content');
     return this;
   }
