+++ package.json
@@ -37,7 +37,8 @@
     "through": "~2.3.4",
     "query-param-getter": "0.0.1",
     "browserify-shim": "~3.3.2",
-    "displayer-api-client": "~0.1.2"
+    "displayer-api-client": "~0.1.2",
+    "openbadges-directory-client": "git://github.com/jpcamara/openbadges-directory-client.git"
   },
   "devDependencies": {
     "mocha": "~1.15.1",
